+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added gcov options and configs to support mcmgr code coverage
-- Addws new test_weak_mu_isr testcase for devices with MU peripheral
+- Added new test_weak_mu_isr testcase for devices with MU peripheral
+- Added new test_heartbeat testcase showing heartbeat mechanism between primary and secondary cores using the MCMGR
 
 ## [v5.0.1]